Change value type to float

+++ src/utils/searchspaces.py
@@ -673,19 +673,19 @@ def create_pca_search_space():
         {
             "name": "variance_e",
             "values": parameter["variance_choices"],
-            "value_type": "int",
+            "value_type": "float",
             "type": "choice",
         },
         {
             "name": "variance_m",
             "values": parameter["variance_choices"],
-            "value_type": "int",
+            "value_type": "float",
             "type": "choice",
         },
         {
             "name": "variance_c",
             "values": parameter["variance_choices"],
-            "value_type": "int",
+            "value_type": "float",
             "type": "choice",
         },
         {
